refactor(docker): move builder out of service to dedicated class
