+++ packages/backend/src/modules/docker/builders/__tests__/compose.builder.test.ts
@@ -1,5 +1,6 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 import { DockerComposeBuilder } from '../compose.builder';
+import type { ServiceInput } from '../schemas';
 import { ServiceBuilder } from '../service.builder';
 
 describe('DockerComposeBuilder', () => {
@@ -117,5 +118,66 @@ describe('DockerComposeBuilder', () => {
     const compose = composeBuilder.addService(service1).addService(service2).addNetwork({ external: true, name: 'network' }).build();
 
     expect(compose).toMatchSnapshot();
+  });
+
+  it('should be able to parse a compose.json file', async () => {
+    const composeJson: { services: ServiceInput[] } = {
+      services: [
+        {
+          // @ts-expect-error
+          something: 'crazy',
+          name: 'ctfd',
+          image: 'ctfd/ctfd:3.7.5',
+          isMain: true,
+          internalPort: 8000,
+          environment: {
+            UPLOAD_FOLDER: '/var/uploads',
+            DATABASE_URL: 'mysql+pymysql://tipi:${CTFD_MYSQL_DB_PASSWORD}@ctfd-db/ctfd',
+          },
+          dependsOn: ['ctfd-db'],
+          volumes: [
+            {
+              hostPath: '${APP_DATA_DIR}/data/uploads',
+              containerPath: '/var/log/CTFd',
+            },
+            {
+              hostPath: '${APP_DATA_DIR}/data/uploads',
+              containerPath: '/var/uploads',
+            },
+          ],
+        },
+        {
+          name: 'ctfd-db',
+          image: 'mariadb:10.4.12',
+          environment: {
+            MYSQL_ROOT_PASSWORD: '${CTFD_MYSQL_ROOT_PASSWORD}',
+            MYSQL_USER: 'tipi',
+            MYSQL_PASSWORD: '${CTFD_MYSQL_DB_PASSWORD}',
+            MYSQL_DATABASE: 'ctfd',
+          },
+          volumes: [
+            {
+              hostPath: '${APP_DATA_DIR}/data/db',
+              containerPath: '/var/lib/mysql',
+            },
+          ],
+          command: ['mysqld', '--character-set-server=utf8mb4', '--collation-server=utf8mb4_unicode_ci', '--wait_timeout=28800', '--log-warnings=0'],
+        },
+        {
+          name: 'ctfd-redis',
+          image: 'redis:4',
+          volumes: [
+            {
+              hostPath: '${APP_DATA_DIR}/data/redis',
+              containerPath: '/data',
+            },
+          ],
+        },
+      ],
+    };
+
+    const yaml = composeBuilder.getDockerCompose(composeJson.services, { appId: 'test-app', openPort: true });
+
+    expect(yaml).toMatchSnapshot();
   });
 });
